backend/requirements-windows.txt
fastapi==0.109.0
fastapi-users[sqlalchemy]==12.1.3
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pyodbc==5.0.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
python-dotenv==1.0.0
email-validator==2.1.0
reportlab==4.0.9
# pandas excluded to prefer prebuilt wheel already installed
# pandas==2.1.4
openpyxl==3.1.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.32.5
